Add warning/error icons for exited resources
