+++ src/Aspire.Dashboard/Components/Pages/Resources.razor
@@ -54,10 +54,7 @@
                         <ResourceNameDisplay Resource="context" FilterText="@_filter" FormatName="GetResourceName" />
                     </TemplateColumn>
                     <TemplateColumn Title="@Loc[Dashboard.Resources.Resources.ResourcesEnvironmentVariablesGridStateColumnHeader]" Sortable="true" SortBy="@_stateSort">
-                        <div class="resource-state-container">
-                            @context.Status?.State
-                            <UnreadLogErrorsBadge UnviewedCount="@GetUnviewedErrorCount(context)" OnClick="@(() => ViewErrorStructuredLogs(context))" />
-                        </div>
+                        <StateColumnDisplay Resource="@context" UnviewedErrorCounts="@_applicationUnviewedErrorCounts" />
                     </TemplateColumn>
                     <PropertyColumn Property="@(c => c.CreationTimeStamp)" Title="@Loc[Dashboard.Resources.Resources.ResourcesEnvironmentVariablesGridStartTimeColumnHeader]" Sortable="true" Tooltip="true" />
                     <TemplateColumn Title="@Loc[Dashboard.Resources.Resources.ResourcesEnvironmentVariablesGridSourceColumnHeader]">
