+++ src/Aspire.Dashboard/Resources/xlf/Columns.de.xlf
@@ -57,6 +57,16 @@
         <target state="new">Working Dir: {0}</target>
         <note>{0} is a path</note>
       </trans-unit>
+      <trans-unit id="StateColumnResourceExited">
+        <source>Resource exited</source>
+        <target state="new">Resource exited</target>
+        <note />
+      </trans-unit>
+      <trans-unit id="StateColumnResourceExitedUnexpectedly">
+        <source>Resource exited unexpectedly with exit code {0}</source>
+        <target state="new">Resource exited unexpectedly with exit code {0}</target>
+        <note>{0} is a number</note>
+      </trans-unit>
       <trans-unit id="UnreadLogErrorsBadgeErrorLog">
         <source>{0} error log</source>
         <target state="new">{0} error log</target>
