+++ src/Aspire.Dashboard/wwwroot/css/app.css
@@ -29,6 +29,7 @@ fluent-toolbar[orientation=horizontal] {
     --log-error: #F8ECEB;
     --log-warning: #FBF3DD;
     --badge-fill-error: var(--error);
+    --badge-fill-warning: var(--warning);
     --badge-color-error: var(--foreground-on-accent-rest);
     --foreground-subtext-rest: #6F6F6F;
     --foreground-subtext-hover: #6C6C6C;
